@@ -2,7 +2,7 @@ import { Injectable } from "@angular/core";
 import { ShoppingCartService } from "../restaurant-detail/shopping-cart/shopping-cart.service";
 import { CartItem } from "../restaurant-detail/shopping-cart/cart-item.model";
 import { Order } from "./order.model";
-import { Http, Headers, RequestOptions } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
 import { MEAT_API } from '../app.api';
@@ -10,7 +10,7 @@ import { MEAT_API } from '../app.api';
 @Injectable()
 export class OrderService {
 
-    constructor(private cartService: ShoppingCartService, private meuHttp: Http) { }
+    constructor(private cartService: ShoppingCartService, private meuHttp: HttpClient) { }
 
     itemsValue(): number {
         return this.cartService.total();
@@ -33,13 +33,7 @@ export class OrderService {
     }
 
     checkOrder(order: Order): Observable<string> {
-        const myheaders = new Headers();
-        myheaders.append('Content-Type', 'application/json');
-
-        return this.meuHttp.post(`${MEAT_API}/orders`,
-            JSON.stringify(order),
-            new RequestOptions({ headers: myheaders }))
-            .map(response => response.json())
+        return this.meuHttp.post<Order>(`${MEAT_API}/orders`, order)
             .map(ordr => ordr.id);
     }
 
