@@ -1,6 +1,6 @@
 import { IRestaurant } from './restaurant/restaurant.model';
 import { Injectable } from '@angular/core';
-import { Http } from '@angular/http';
+import { HttpClient, HttpParams } from '@angular/common/http';
 import { MEAT_API } from '../app.api';
 import { Observable } from '../../../node_modules/rxjs/Observable';
 import 'rxjs/add/operator/map';
@@ -12,29 +12,29 @@ import { MenuItem } from '../restaurant-detail/menu-item/menu-item.model';
 @Injectable()
 export class RestaurantsService {
 
-    constructor(private myHttp: Http) { }
+    constructor(private myHttp: HttpClient) { }
 
     myRestaurants: IRestaurant[] = []
 
 
     GetMyRestaurants(search?: string): Observable<IRestaurant[]> {
-        return this.myHttp.get(`${MEAT_API}/restaurants`, { params: { q: search } })
-            .map(response => response.json());
+        let params: HttpParams = undefined;
+        if (search) {
+            params = new HttpParams().append('q', search);
+        }
+        return this.myHttp.get<IRestaurant[]>(`${MEAT_API}/restaurants`, { params: params });
     }
 
     RestaurantsById(id: string): Observable<IRestaurant> {
 
-        return this.myHttp.get(`${MEAT_API}/restaurants/${id}`)
-            .map(Response => Response.json());
+        return this.myHttp.get<IRestaurant>(`${MEAT_API}/restaurants/${id}`)
     }
 
     ReviewsOfRestaurant(id: string): Observable<any> {
         return this.myHttp.get(`${MEAT_API}/restaurants/${id}/reviews`)
-            .map(Response => Response.json());
     }
 
     MenuOfRestaurant(id: string): Observable<MenuItem[]> {
-        return this.myHttp.get(`${MEAT_API}/restaurants/${id}/menu`)
-            .map(Response => Response.json());
+        return this.myHttp.get<MenuItem[]>(`${MEAT_API}/restaurants/${id}/menu`)
     }
 }
